+++ src/main/java/view/InputView.java
@@ -1,17 +1,14 @@
 package view;
 
-import domain.Attendance;
-import domain.Calender;
-import java.time.LocalDate;
 import java.util.regex.Pattern;
 import util.Console;
-import util.FileManager;
-import util.TodayDateTimeUtil;
+import util.DateTimeUtil;
 
 public class InputView {
     public static String inputFeature() {
         System.out.printf("오늘은 %d월 %02d일 %s입니다. 기능을 선택해 주세요.\n" + "1. 출석 확인\n" + "2. 출석 수정\n" + "3. 크루별 출석 기록 확인\n"
-                + "4. 제적 위험자 확인\n" + "Q. 종료\n", TodayDateTimeUtil.nowMonth(),  TodayDateTimeUtil.nowDate(),  TodayDateTimeUtil.nowDayOfWeek());
+                        + "4. 제적 위험자 확인\n" + "Q. 종료\n", DateTimeUtil.nowMonth(), DateTimeUtil.nowDate(),
+                DateTimeUtil.nowDayOfWeek());
         String featureNumber = Console.readLine();
         validateFeatureNumber(featureNumber);
         return featureNumber;
